BACKPORT: Merge git://git.kernel.org/pub/scm/linux/kernel/git/netdev/net
xdp_umem.c had overlapping changes between the 64-bit math fix
for the calculation of npgs and the removal of the zerocopy
memory type which got rid of the chunk_size_nohdr member.

The mlx5 Kconfig conflict is a case where we just take the
net-next copy of the Kconfig entry dependency as it takes on
the ESWITCH dependency by one level of indirection which is
what the 'net' conflicting change is trying to ensure.

Signed-off-by: David S. Miller <davem@davemloft.net>

[auto-resolved: BPF-core=INCOMING, non-BPF=OURS]
(cherry picked from commit 1806c13dc2532090d742ce03847b22367fb20ad6)
Change-Id: I6e50b4f1004309ac841b2d034029206d0a0f5eed