modpost: pass -N option only for modules modpost
The built-in only code is not required to have MODULE_IMPORT_NS() to
use symbols. So, the namespace is not checked for vmlinux(.o).

Do not pass the meaningless -N option to the first pass of modpost.

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
(cherry picked from commit 4e5ab74c3cbbe7ca2b907a86ce5140e442b340bf)