BACKPORT: bpf: Implement bpf_local_storage for inodes
Similar to bpf_local_storage for sockets, add local storage for inodes.
The life-cycle of storage is managed with the life-cycle of the inode.
i.e. the storage is destroyed along with the owning inode.

The BPF LSM allocates an __rcu pointer to the bpf_local_storage in the
security blob which are now stackable and can co-exist with other LSMs.

Signed-off-by: KP Singh <kpsingh@google.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Link: https://lore.kernel.org/bpf/20200825182919.1118197-6-kpsingh@chromium.org

[auto-resolved: BPF-core=INCOMING, non-BPF=OURS]
(cherry picked from commit 8ea636848aca35b9f97c5b5dee30225cf2dd0fe6)
Change-Id: Ib402b51d972ccd4e38c8a98aa768154bf5c7b34d